Move RootNode.
